move functions to derive node predictions to parser.py and make changes to work with renamed columns
